consolidate progress components in one directory
