+++ src/module/vue/components/completed-progress-list.vue
@@ -21,9 +21,9 @@ import type { ExtractPropTypes, Ref } from 'vue'
 import { computed, inject, ref, watch } from 'vue'
 import { ActorKey } from '../provisions'
 import Collapsible from './collapsible/collapsible.vue'
-import type { CompletedProgressType } from './progress-common'
-import { getProgressItems } from './progress-common'
-import ProgressList from './progress-list.vue'
+import type { CompletedProgressType } from './progress/progress-common'
+import { getProgressItems } from './progress/progress-common'
+import ProgressList from './progress/progress-list.vue'
 
 const props = defineProps<{
 	collapsibleProps?: Omit<
