+++ src/module/vue/components/progress/progress-controls.vue
