+++ src/module/vue/components/progress/progress-item-detail.vue
